Merge in DTX/relativitysync from develop to master

* commit 'e6938853742a3c9a4a57d93af1589f2c6e1bee55':
  REL-699100 Add Retry around Bad Gateway exception
  [REL-690539] ConfigureAwait(false) added to async methods invokes in ARMHelper
  [REL-690539] ARMHelper cleanup
  [REL-690539] ARM Run and Status check moved into one retry policy
  [REL-690539] ARM Status check retries number increased to 5
  [REL-690539] Retry Policy applied to ARM Job status state check
  [REL-690539] Retry Policy added to ARM Restore Run
  REL-686367 Added caching to SyncToggles
  REL-679803 Upgrade IAPI to 1.17.14
  REL-686367 Fixed toggle description
  REL-686375 Read UseFMS toggle in ImportJobFactory
  REL-686367 Added UseFMS toggle and refactored toggle service
  REL-690552 Updated issues.json
  REL-682915 Fixed typo
  REL-682915 Review remarks
  REL-682915 Added ADR - RAP Workflow and Architecture
  REL-680845 Clean SyncBatches after Sync Job finish (Drain-Stop is exception)
  REL-661414 Add additional logging for Drain-Stop feature

+++ Source/Relativity.Sync/Executors/JobCleanupExecutor.cs
@@ -21,6 +21,11 @@ namespace Relativity.Sync.Executors
 		{
 			try
 			{
+				if(token.IsDrainStopRequested)
+				{
+					return ExecutionResult.Success();
+				}
+
 				await _batchRepository.DeleteAllForConfigurationAsync(configuration.SourceWorkspaceArtifactId, configuration.SyncConfigurationArtifactId).ConfigureAwait(false);
 				return ExecutionResult.Success();
 			}
